Enable FIX trading, move login details to ibc's config.ini
- IBC supports CTCI FIX login, but ib-gateway-docker does
not currently provide a method to use it by default. I had issues
passing the FIX login details to IBC via the CLI, so I opted for
the config.ini.tmpl approach as used with the READ_ONLY_API variable,
and it worked without issue. I then migrated the TWS login variables to
that approach too for consistency.

- jts.ini has been updated to reflect FIX-related settings. I do
not know which of these are mandatory and which are too specific for
general use cases, so feedback is welcome. I do not believe this update
will cause issues for accounts that are not FIX enabled, but this will
need verification.

- fork_ports_delayed.sh has been updated to expose port 4010 as the FIX
communication port. It maps to port 4020 in the container, which is also
set as the FIX port in jts.ini. I wonder if these should instead be dictated by
environment variables.

- The README and docker-compose.yml have been updated to demonstrate
FIX login.